Unificate browser and ssr components
